Move naming contracts into shared
