+++ src/kernel/naming/query.ts
@@ -16,6 +16,9 @@
 
 import type { OcctBackend } from '../backends/occt/occtBackend';
 import type { FeatureRecord } from '../../shared/intent/featureRecord';
+import type { QueryAst, QueryKind } from '../../shared/naming/queryAst';
+
+export type { GeometryType, QueryAst, QueryKind } from '../../shared/naming/queryAst';
 
 /** Internal: evaluator delegates injected by `queryEvaluator.ts` after its
  *  module init finishes. The chainables read this table at call time so the
@@ -46,9 +49,6 @@ export function __installQueryStringifier(fn: (q: Query<unknown>) => string): vo
   __formatQueryAsString = fn;
 }
 
-/** Discriminator for the entity kind a Query addresses. */
-export type QueryKind = 'face' | 'edge' | 'vertex' | 'connector' | 'part' | 'solid';
-
 /** Phantom type markers for static type narrowing. The markers are NEVER
  *  materialised at runtime; the runtime discriminator is `Query.target`.
  *  Phantoms exist purely so TS rejects `fillet({ edges: faceQuery })` at
@@ -63,29 +63,6 @@ export type SolidMarker = { readonly _solid: unique symbol };
 export type EntityMarker =
   | FaceMarker | EdgeMarker | VertexMarker
   | ConnectorMarker | PartMarker | SolidMarker;
-
-export type GeometryType =
-  | 'PLANE' | 'CYLINDER' | 'CONE' | 'SPHERE' | 'TORUS'
-  | 'BSPLINE_SURFACE' | 'LINE' | 'CIRCLE' | 'BSPLINE_CURVE' | 'OTHER';
-
-/** The Query AST. Serializable; structurally equal under round-trip. */
-export type QueryAst =
-  | { op: 'nothing' }
-  | { op: 'everything'; kind: QueryKind }
-  | { op: 'createdBy'; id: string; kind?: QueryKind }
-  | { op: 'ownedByPart'; query: QueryAst }
-  | { op: 'ownerPart'; query: QueryAst }
-  | { op: 'union'; queries: QueryAst[] }
-  | { op: 'intersection'; queries: QueryAst[] }
-  | { op: 'subtraction'; a: QueryAst; b: QueryAst }
-  | { op: 'containsPoint'; query: QueryAst; point: [number, number, number] }
-  | { op: 'closestTo'; query: QueryAst; point: [number, number, number]; k?: number }
-  | { op: 'geometryType'; query: QueryAst; geomType: GeometryType }
-  | { op: 'entityFilter'; query: QueryAst; kind: QueryKind }
-  | { op: 'withLabel'; query: QueryAst; label: string }
-  | { op: 'withFeatureName'; query: QueryAst; name: string }
-  | { op: 'nthElement'; query: QueryAst; index: number }
-  | { op: 'fromString'; ref: string };
 
 /** A resolved entity carries the OCCT handle, the canonical @kc[...] ref
  *  (so diagnostic prose stays compact), and a snapshot for cite-by-geometry. */
